Web UI on Android Smartphones or Tablets: next/previous Station Button switches two stations forward or back #286
On systems with mouse operation there are the html events "onmousedown" and "onmouseup", on platforms like Android "ontouchstart" and "ontouching". Probably on newer devices/versions for touch operation the mouse events are also emulated. On my tablet I was able to reproduce the error. I have now used "onclick" for previous and next station.